restructure tests and fix typos
