+++ packages/xlucene-translator/test/structure/cases/translator/arrays.ts
