+++ packages/xlucene-translator/test/structure/cases/translator/interfaces.ts
@@ -1,7 +1,7 @@
 import {
     AnyQuery, ElasticsearchDSLResult, ElasticsearchDSLOptions
 } from '@terascope/types';
-import { TranslatorOptions } from '../../../src/translator/interfaces.js';
+import { TranslatorOptions } from '../../../../src/translator/interfaces.js';
 
 export type TestCase = [
     // when given %s
