+++ packages/xlucene-translator/test/structure/cases/translator/logical-group.ts
