+++ packages/xlucene-translator/test/structure/cases/translator/term.ts
